feat(delegation): auto associate when possible (#284)
* feat(delegation): auto associate when possible

Whenever the asset bearing `ExocoreAssetID` is delegated by a staker to
an operator, we should check if the staker address is the same as the
operator address. If a match is found, they should be automatically
associated.

* doc(delegation): update comment about duplicate

- It is not possible to overwrite an association of the `stakerID`
  ending with `0x0` because such an association cannot be made using any
  other mechanism.
- Even if it exists, the overwriting will do nothing of import due to
  the equality check.

* test(delegation): check operator share auto assoc

Validate that the operator share is correctly set to the expected value
when self-staking ExocoreAssetID

* doc(assets): clarify StakerAssetInfo

- for ExocoreAssetID, the `TotalDepositAmount` is post-slashing, if any
  was applied.
- for other asset IDs, all components are pre-slashing.